+++ CHANGELOG.md
@@ -12,6 +12,7 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 - Georeferences coordinates in label generate on New collecting event [#2107]
 - Lock buttons on New image [#2101]
 - Open PDF slider in all tabs [#2106]
+- Preparation type and with/out facets [#2118]
 
 ### Changed
 - Nexus output file was modified to present full name of the of the taxon. TNT export was not changed.
@@ -24,6 +25,7 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 [#2106]: https://github.com/SpeciesFileGroup/taxonworks/issues/2101
 [#2107]: https://github.com/SpeciesFileGroup/taxonworks/issues/2107
 [#2112]: https://github.com/SpeciesFileGroup/taxonworks/issues/2112
+[#2118]: https://github.com/SpeciesFileGroup/taxonworks/issues/2118
 
 ## [0.16.4] - 2020-03-09